Eslint & Cleanup
