+++ packages/components/sections/tooltips/TooltipsPersonalBlock.tsx
@@ -6,9 +6,9 @@ import { useIndexedDB } from '~/hooks/useIndexedDB';
 import type { CustomItem, Tooltip } from '~/types/item';
 import { useAppDispatch, useAppSelector } from '~/lib/store/hooks';
 import type { RootState } from '~/lib/store/store';
-import { selectItemById, setIsCustomTooltipsOpen, setItem, setSelectedItem } from '~/lib/store/features/appSlice';
+import { selectItemById, setIsCustomTooltipsOpen } from '~/lib/store/features/appSlice';
 import { useConfirmation } from '~/providers/ConfirmationProvider';
-import { useShortcut } from '~/hooks/useShortcut';
+import { useFileOperations } from '~/hooks/useFileOperations';
 import { useItem } from '~/hooks/useItem';
 
 interface TooltipsPersonalBlockProps {
@@ -26,7 +26,7 @@ const TooltipsPersonalBlock: React.FC<TooltipsPersonalBlockProps> = ({
     const { updateItem } = useItem();
     const { requestConfirmation } = useConfirmation();
     const { getTooltipConfigurations, deleteTooltipConfiguration } = useIndexedDB();
-    const { exportTooltipConfiguration, importTooltipConfiguration } = useShortcut();
+    const { exportTooltipConfiguration, importTooltipConfiguration } = useFileOperations();
 
     const selectedItem = useAppSelector((state: RootState) => state.app.selectedItem);
     const currentItem = useAppSelector((state: RootState) => selectItemById(state, selectedItem ?? ''));
